perf: avoid eager CLI imports for fast paths
