+++ src/main.js
@@ -31,7 +31,8 @@ import { SUPPORTED_PROVIDERS, assertSupportedProvider, buildProviderTemplateComm
 import { runRepoSync } from "./core/repo-sync.js";
 import { buildSkillInstallHint, installAllBuiltinSkills, installBuiltinSkill, listBuiltinSkills } from "./core/skills.js";
 import { runBootstrapCommand } from "./core/bootstrap.js";
-import { VERSION, withSilent, bold, cyan, dim, green, success, log } from "./core/ui.js";
+import { VERSION, printHelp } from "./core/cli-fast-paths.js";
+import { withSilent, bold, dim, green, success, log } from "./core/ui.js";
 
 function parseValue(raw) {
   if (raw === "true") {
@@ -379,107 +380,6 @@ async function maybePrintPreparedChild(root, config, launch) {
   return child;
 }
 
-function printHelp() {
-  const c = (s) => bold(cyan(s));
-  const d = (s) => dim(s);
-
-  const lines = [
-    `  ${bold("COMMANDS")}`,
-    ``,
-    `    ${c("init")}            ${d("Create baseline Agentify artifacts")}`,
-    `    ${c("index")}           ${d("Build the SQLite repository index")}`,
-    `    ${c("scan")}            ${d("Alias for index")}`,
-    `    ${c("doc")}             ${d("Generate docs, metadata, and key-file headers")}`,
-    `    ${c("up")}              ${d("Run scan -> optional doc -> check -> test pipeline")}`,
-    `    ${c("sync")}            ${d("Upgrade repo-owned Agentify files, then run refresh")}`,
-    `    ${c("check")}           ${d("Validate freshness, schemas, and safety rules")}`,
-    `    ${c("plan")}            ${d("Preview the planner-selected context for a task")}`,
-    `    ${c("context")}         ${d("Search indexed context and fetch exact bounded file slices")}`,
-    `    ${c("run")}             ${d("Run provider template command with auto-refresh")}`,
-    `    ${c("exec")}            ${d("Advanced wrapper for custom agent commands")}`,
-    `    ${c("this")}            ${d("Bootstrap this macOS repo for a provider-backed Agentify workflow")}`,
-    `    ${c("context")}         ${d("Search, fetch, compact, and inspect routed context")}`,
-    `    ${c("query")}           ${d("Query the repository index (owner, deps, changed, def, refs, callers, impacts)")}`,
-    `    ${c("risk")}            ${d("Score PR blast radius and recommend regression tests")}`,
-    `    ${c("skill")}           ${d("Manage built-in agent skills")}`,
-    `    ${c("sess")}            ${d("Manage provider-backed sessions")}`,
-    `    ${c("handoff")}         ${d("Write a cross-agent handoff bundle for a session")}`,
-    `    ${c("memory")}          ${d("Manage agent memory helpers")}`,
-    `    ${c("issue-killer")}    ${d("Launch labelled GitHub issues into supervised tmux worktrees")}`,
-    `    ${c("hooks")}           ${d("Install/remove git hooks")}`,
-    `    ${c("doctor")}          ${d("Check toolchain health and capability tier")}`,
-    `    ${c("semantic")}        ${d("Refresh semantic project facts")}`,
-    `    ${c("clean")}           ${d("Prune stale generated artifacts and dead Agentify folders")}`,
-    `    ${c("cache")}           ${d("Manage the content cache")}`,
-    ``,
-    `  ${bold("OPTIONS")}`,
-    ``,
-    `    ${c("--provider")} ${d(`<${SUPPORTED_PROVIDERS.join("|")}>`)}`,
-    `                         ${d("skill install also accepts comma lists and all")}`,
-    `    ${c("--strict")} ${d("<true|false>")}         Fail closed on validation issues`,
-    `    ${c("--languages")} ${d("<auto|ts|python|go|rust|dotnet|java|kotlin|swift>")}`,
-    `    ${c("--dry-run")}                   Report planned changes without writing`,
-    `    ${c("--docs")}                      Generate docs during refresh/update flows (on by default; use --docs=false to skip)`,
-    `    ${c("--headers")}                   Apply @agentify headers to source files (off by default)`,
-    `    ${c("--semantic")}                  Show detailed semantic diagnostics with doctor`,
-    `    ${c("--provider-timeout-ms")} ${d("<ms>")}     Fail provider doc calls after N milliseconds`,
-    `    ${c("--ghost")}                     Route outputs to .current_session/`,
-    `    ${c("--json")}                      Machine-readable JSON output only`,
-    `    ${c("--explain")}                   Include planner score breakdowns for plan output`,
-    `    ${c("--interactive")}, ${c("-i")}       Force interactive mode (template providers default to interactive for run/sess)`,
-    `    ${c("--continue")}                  Resume the provider's most recent session for run`,
-    `    ${c("--resume")}                    Alias for run --continue; with session/sess, resume Agentify session context`,
-    `    ${c("--context-mode")} ${d("<compact|routed>")}  Use compact prompts or routed bounded retrieval prompts`,
-    `    ${c("--with-context")}              Inject planner-selected files, tests, and memory into run`,
-    `    ${c("--context-mode")} ${d("<direct|routed>")}     Use routed context retrieval for run/sess prompts`,
-    `    ${c("--bypass-permissions")}        Explicitly bypass provider permission prompts for issue-killer panes`,
-    `    ${c("--explain-plan")}              Print planner output before executing run`,
-    `    ${c("--caveman[=level]")}            Terse output for run/sess (lite, full, ultra, wenyan*)`,
-    `    ${c("--root")} ${d("<path>")}               Target repo root (default: cwd)`,
-    `    ${c("--scope")} ${d("<project|user>")}      Skill install scope (skill command)`,
-    `    ${c("--hook")}                      Hook-friendly validation for check/up: skip source body diffing`,
-    ``,
-    `  ${bold("EXEC FLAGS")}`,
-    ``,
-    `    ${c("--fail-on-stale")}             Exit 80 if validation fails post-refresh`,
-    `    ${c("--timeout")} ${d("<seconds>")}         Kill wrapped command after N seconds`,
-    `    ${c("--skip-refresh")}              Skip post-command refresh`,
-    ``,
-    `  ${bold("EXAMPLES")}`,
-    ``,
-    `    ${d("$")} agentify init`,
-    `    ${d("$")} agentify this --provider codex`,
-    `    ${d("$")} agentify up --provider codex`,
-    `    ${d("$")} agentify sync`,
-    `    ${d("$")} agentify clean --dry-run`,
-    `    ${d("$")} agentify run --provider codex`,
-    `    ${d("$")} agentify run --provider codex "implement payment retries"`,
-    `    ${d("$")} agentify run --provider codex --resume`,
-    `    ${d("$")} agentify run --provider codex --caveman=ultra "summarize auth risks"`,
-    `    ${d("$")} agentify run --provider codex --interactive "fix auth bug"`,
-    `    ${d("$")} agentify context search analytics`,
-    `    ${d("$")} agentify context fetch src/analytics/report.ts --symbol buildReport`,
-    `    ${d("$")} agentify context fetch src/analytics/report.ts --lines 20:60`,
-    `    ${d("$")} agentify risk --since origin/main`,
-    `    ${d("$")} agentify risk --json`,
-    `    ${d("$")} agentify skill list`,
-    `    ${d("$")} agentify skill install all --provider codex --scope project`,
-    `    ${d("$")} agentify skill install grill-me --provider claude --scope project`,
-    `    ${d("$")} agentify skill install god-mode --provider all --scope project`,
-    `    ${d("$")} agentify memory compress AGENTIFY.md`,
-    `    ${d("$")} agentify sess run --provider codex --name "payments-v2"`,
-    `    ${d("$")} agentify sess run --provider codex --name "payments-v2" "add tests"`,
-    `    ${d("$")} agentify sess run --provider codex --interactive --name "payments-v2" "continue in Codex TUI"`,
-    `    ${d("$")} agentify handoff --session sess_20260101000000_abcdef "continue payments-v2"`,
-    `    ${d("$")} agentify issue-killer --label agentify-ready --agent-provider codex --limit 5`,
-    `    ${d("$")} agentify issue-killer --label agentify-ready --agent-provider codex --bypass-permissions`,
-    `    ${d("$")} agentify exec -- codex exec "fix auth bug"`,
-    ``,
-  ];
-
-  process.stderr.write(lines.join("\n") + "\n");
-}
-
 export function parseArgs(argv) {
   const args = { _: [] };
   let seenDoubleDash = false;
@@ -560,7 +460,7 @@ export async function runCli(argv, runtime = {}) {
   }
 
   if (command === "help" || args.help) {
-    printHelp();
+    await printHelp();
     return;
   }
